+++ test/completion/init.lua
@@ -1784,7 +1784,6 @@ zz$
 global zzz: integer = 1
 ```
 ---
-
  abc]],
     }
 }