+++ changelog.md
@@ -5,6 +5,7 @@
 * `FIX` completion: interface in nested table
 * `FIX` completion: interface not show after `,`
 * `FIX` [#400](https://github.com/sumneko/lua-language-server/issues/400)
+* `FIX` [#402](https://github.com/sumneko/lua-language-server/issues/402)
 * `FIX` runtime errors
 
 ## 1.16.0