+++ changelog.md
@@ -7,6 +7,7 @@
 * `FIX` [#400](https://github.com/sumneko/lua-language-server/issues/400)
 * `FIX` [#402](https://github.com/sumneko/lua-language-server/issues/402)
 * `FIX` [#403](https://github.com/sumneko/lua-language-server/issues/403)
+* `FIX` [#404](https://github.com/sumneko/lua-language-server/issues/404)
 * `FIX` runtime errors
 
 ## 1.16.0